simple-dapr-server/requirements.txt
flask
dapr-ext-grpc-dev >= 1.0.0rc3.dev
dapr-dev >= 1.0.0rc3.dev